Wywalenie wszystkich glPush, glPop, glTranslate i glScaled. Zepsucie tekstu, bo on nie używa shaderów - renderuje się w złym miejscu. Jedno glOrtho jeszcze zostało, ale tylko po to, żeby tekst mógł się dobrze renderować.

+++ src/game/place/cameras/Camera.java
@@ -312,8 +312,8 @@ public abstract class Camera {
         this.delayLength = delayLength;
     }
 
-    public double getScale() {
-        return scale;
+    public float getScale() {
+        return (float) scale;
     }
 
     public int getArea() {
